requirements.txt
transformers
pyannote.audio
python-docx
